split phue module into a package
