+++ phue/__init__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+
+'''
+phue by Nathanaël Lécaudé - A Philips Hue Python library
+Contributions by Marshall Perrin, Justin Lintz
+https://github.com/studioimaginaire/phue
+Original protocol hacking by rsmck : http://rsmck.co.uk/hue
+
+Published under the MIT license - See LICENSE file for more details.
+
+"Hue Personal Wireless Lighting" is a trademark owned by Koninklijke Philips Electronics N.V., see www.meethue.com for more information.
+I am in no way affiliated with the Philips organization.
+
+'''
+
+__version__ = '0.9'
