+++ phue/bridge.py
@@ -1,577 +1,17 @@
-#!/usr/bin/python
 # -*- coding: utf-8 -*-
 
-'''
-phue by Nathanaël Lécaudé - A Philips Hue Python library
-Contributions by Marshall Perrin, Justin Lintz
-https://github.com/studioimaginaire/phue
-Original protocol hacking by rsmck : http://rsmck.co.uk/hue
-
-Published under the MIT license - See LICENSE file for more details.
-
-"Hue Personal Wireless Lighting" is a trademark owned by Koninklijke Philips Electronics N.V., see www.meethue.com for more information.
-I am in no way affiliated with the Philips organization.
-
-'''
-
-import json
-import logging
 import os
-import platform
-import sys
+import json
 import socket
-if sys.version_info[0] > 2:
-    PY3K = True
-else:
-    PY3K = False
+
+from .utils import USER_HOME, MOBILE, PY3K, logger, is_string
+from .exception import PhueException, PhueRegistrationException, PhueRequestTimeout
+from .light import Light
 
 if PY3K:
     import http.client as httplib
 else:
     import httplib
-
-logger = logging.getLogger('phue')
-
-
-if platform.system() == 'Windows':
-    USER_HOME = 'USERPROFILE'
-else:
-    USER_HOME = 'HOME'
-
-__version__ = '0.9'
-
-
-def is_string(data):
-    """Utility method to see if data is a string."""
-    if PY3K:
-        return isinstance(data, str)
-    else:
-        return isinstance(data, str) or isinstance(data, unicode)  # noqa
-
-
-class PhueException(Exception):
-
-    def __init__(self, id, message):
-        self.id = id
-        self.message = message
-
-
-class PhueRegistrationException(PhueException):
-    pass
-
-
-class PhueRequestTimeout(PhueException):
-    pass
-
-
-class Light(object):
-
-    """ Hue Light object
-
-    Light settings can be accessed or set via the properties of this object.
-
-    """
-    def __init__(self, bridge, light_id):
-        self.bridge = bridge
-        self.light_id = light_id
-
-        self._name = None
-        self._on = None
-        self._brightness = None
-        self._colormode = None
-        self._hue = None
-        self._saturation = None
-        self._xy = None
-        self._colortemp = None
-        self._effect = None
-        self._alert = None
-        self.transitiontime = None  # default
-        self._reset_bri_after_on = None
-        self._reachable = None
-        self._type = None
-
-    def __repr__(self):
-        # like default python repr function, but add light name
-        return '<{0}.{1} object "{2}" at {3}>'.format(
-            self.__class__.__module__,
-            self.__class__.__name__,
-            self.name,
-            hex(id(self)))
-
-    # Wrapper functions for get/set through the bridge, adding support for
-    # remembering the transitiontime parameter if the user has set it
-    def _get(self, *args, **kwargs):
-        return self.bridge.get_light(self.light_id, *args, **kwargs)
-
-    def _set(self, *args, **kwargs):
-
-        if self.transitiontime is not None:
-            kwargs['transitiontime'] = self.transitiontime
-            logger.debug("Setting with transitiontime = {0} ds = {1} s".format(
-                self.transitiontime, float(self.transitiontime) / 10))
-
-            if (args[0] == 'on' and args[1] is False) or (
-                    kwargs.get('on', True) is False):
-                self._reset_bri_after_on = True
-        return self.bridge.set_light(self.light_id, *args, **kwargs)
-
-    @property
-    def name(self):
-        '''Get or set the name of the light [string]'''
-        if PY3K:
-            self._name = self._get('name')
-        else:
-            self._name = self._get('name').encode('utf-8')
-        return self._name
-
-    @name.setter
-    def name(self, value):
-        old_name = self.name
-        self._name = value
-        self._set('name', self._name)
-
-        logger.debug("Renaming light from '{0}' to '{1}'".format(
-            old_name, value))
-
-        self.bridge.lights_by_name[self.name] = self
-        del self.bridge.lights_by_name[old_name]
-
-    @property
-    def on(self):
-        '''Get or set the state of the light [True|False]'''
-        self._on = self._get('on')
-        return self._on
-
-    @on.setter
-    def on(self, value):
-
-        # Some added code here to work around known bug where
-        # turning off with transitiontime set makes it restart on brightness = 1
-        # see
-        # http://www.everyhue.com/vanilla/discussion/204/bug-with-brightness-when-requesting-ontrue-transitiontime5
-
-        # if we're turning off, save whether this bug in the hardware has been
-        # invoked
-        if self._on and value is False:
-            self._reset_bri_after_on = self.transitiontime is not None
-            if self._reset_bri_after_on:
-                logger.warning(
-                    'Turned off light with transitiontime specified, brightness will be reset on power on')
-
-        self._set('on', value)
-
-        # work around bug by resetting brightness after a power on
-        if self._on is False and value is True:
-            if self._reset_bri_after_on:
-                logger.warning(
-                    'Light was turned off with transitiontime specified, brightness needs to be reset now.')
-                self.brightness = self._brightness
-                self._reset_bri_after_on = False
-
-        self._on = value
-
-    @property
-    def colormode(self):
-        '''Get the color mode of the light [hs|xy|ct]'''
-        self._colormode = self._get('colormode')
-        return self._colormode
-
-    @property
-    def brightness(self):
-        '''Get or set the brightness of the light [0-254].
-
-        0 is not off'''
-
-        self._brightness = self._get('bri')
-        return self._brightness
-
-    @brightness.setter
-    def brightness(self, value):
-        self._brightness = value
-        self._set('bri', self._brightness)
-
-    @property
-    def hue(self):
-        '''Get or set the hue of the light [0-65535]'''
-        self._hue = self._get('hue')
-        return self._hue
-
-    @hue.setter
-    def hue(self, value):
-        self._hue = int(value)
-        self._set('hue', self._hue)
-
-    @property
-    def saturation(self):
-        '''Get or set the saturation of the light [0-254]
-
-        0 = white
-        254 = most saturated
-        '''
-        self._saturation = self._get('sat')
-        return self._saturation
-
-    @saturation.setter
-    def saturation(self, value):
-        self._saturation = value
-        self._set('sat', self._saturation)
-
-    @property
-    def xy(self):
-        '''Get or set the color coordinates of the light [ [0.0-1.0, 0.0-1.0] ]
-
-        This is in a color space similar to CIE 1931 (but not quite identical)
-        '''
-        self._xy = self._get('xy')
-        return self._xy
-
-    @xy.setter
-    def xy(self, value):
-        self._xy = value
-        self._set('xy', self._xy)
-
-    @property
-    def colortemp(self):
-        '''Get or set the color temperature of the light, in units of mireds [154-500]'''
-        self._colortemp = self._get('ct')
-        return self._colortemp
-
-    @colortemp.setter
-    def colortemp(self, value):
-        if value < 154:
-            logger.warn('154 mireds is coolest allowed color temp')
-        elif value > 500:
-            logger.warn('500 mireds is warmest allowed color temp')
-        self._colortemp = value
-        self._set('ct', self._colortemp)
-
-    @property
-    def colortemp_k(self):
-        '''Get or set the color temperature of the light, in units of Kelvin [2000-6500]'''
-        self._colortemp = self._get('ct')
-        return int(round(1e6 / self._colortemp))
-
-    @colortemp_k.setter
-    def colortemp_k(self, value):
-        if value > 6500:
-            logger.warn('6500 K is max allowed color temp')
-            value = 6500
-        elif value < 2000:
-            logger.warn('2000 K is min allowed color temp')
-            value = 2000
-
-        colortemp_mireds = int(round(1e6 / value))
-        logger.debug("{0:d} K is {1} mireds".format(value, colortemp_mireds))
-        self.colortemp = colortemp_mireds
-
-    @property
-    def effect(self):
-        '''Check the effect setting of the light. [none|colorloop]'''
-        self._effect = self._get('effect')
-        return self._effect
-
-    @effect.setter
-    def effect(self, value):
-        self._effect = value
-        self._set('effect', self._effect)
-
-    @property
-    def alert(self):
-        '''Get or set the alert state of the light [select|lselect|none]'''
-        self._alert = self._get('alert')
-        return self._alert
-
-    @alert.setter
-    def alert(self, value):
-        if value is None:
-            value = 'none'
-        self._alert = value
-        self._set('alert', self._alert)
-
-    @property
-    def reachable(self):
-        '''Get the reachable state of the light [boolean]'''
-        self._type = self._get('reachable')
-        return self._type
-
-    @property
-    def type(self):
-        '''Get the type of the light [string]'''
-        self._type = self._get('type')
-        return self._type
-
-
-class SensorState(dict):
-    def __init__(self, bridge, sensor_id):
-        self._bridge = bridge
-        self._sensor_id = sensor_id
-
-    def __setitem__(self, key, value):
-        dict.__setitem__(self, key, value)
-        self._bridge.set_sensor_state(self._sensor_id, self)
-
-
-class SensorConfig(dict):
-    def __init__(self, bridge, sensor_id):
-        self._bridge = bridge
-        self._sensor_id = sensor_id
-
-    def __setitem__(self, key, value):
-        dict.__setitem__(self, key, value)
-        self._bridge.set_sensor_config(self._sensor_id, self)
-
-
-class Sensor(object):
-
-    """ Hue Sensor object
-
-    Sensor config and state can be read and updated via the properties of this object
-
-    """
-    def __init__(self, bridge, sensor_id):
-        self.bridge = bridge
-        self.sensor_id = sensor_id
-
-        self._name = None
-        self._model = None
-        self._swversion = None
-        self._type = None
-        self._uniqueid = None
-        self._manufacturername = None
-        self._state = SensorState(bridge, sensor_id)
-        self._config = {}
-        self._recycle = None
-
-    def __repr__(self):
-        # like default python repr function, but add sensor name
-        return '<{0}.{1} object "{2}" at {3}>'.format(
-            self.__class__.__module__,
-            self.__class__.__name__,
-            self.name,
-            hex(id(self)))
-
-    # Wrapper functions for get/set through the bridge
-    def _get(self, *args, **kwargs):
-        return self.bridge.get_sensor(self.sensor_id, *args, **kwargs)
-
-    def _set(self, *args, **kwargs):
-        return self.bridge.set_sensor(self.sensor_id, *args, **kwargs)
-
-    @property
-    def name(self):
-        '''Get or set the name of the sensor [string]'''
-        if PY3K:
-            self._name = self._get('name')
-        else:
-            self._name = self._get('name').encode('utf-8')
-        return self._name
-
-    @name.setter
-    def name(self, value):
-        old_name = self.name
-        self._name = value
-        self._set('name', self._name)
-
-        logger.debug("Renaming sensor from '{0}' to '{1}'".format(
-            old_name, value))
-
-        self.bridge.sensors_by_name[self.name] = self
-        del self.bridge.sensors_by_name[old_name]
-
-    @property
-    def modelid(self):
-        '''Get a unique identifier of the hardware model of this sensor [string]'''
-        self._modelid = self._get('modelid')
-        return self._modelid
-
-    @property
-    def swversion(self):
-        '''Get the software version identifier of the sensor's firmware [string]'''
-        self._swversion = self._get('swversion')
-        return self._swversion
-
-    @property
-    def type(self):
-        '''Get the sensor type of this device [string]'''
-        self._type = self._get('type')
-        return self._type
-
-    @property
-    def uniqueid(self):
-        '''Get the unique device ID of this sensor [string]'''
-        self._uniqueid = self._get('uniqueid')
-        return self._uniqueid
-
-    @property
-    def manufacturername(self):
-        '''Get the name of the manufacturer [string]'''
-        self._manufacturername = self._get('manufacturername')
-        return self._manufacturername
-
-    @property
-    def state(self):
-        ''' A dictionary of sensor state. Some values can be updated, some are read-only. [dict]'''
-        data = self._get('state')
-        self._state.clear()
-        self._state.update(data)
-        return self._state
-
-    @state.setter
-    def state(self, data):
-        self._state.clear()
-        self._state.update(data)
-
-    @property
-    def config(self):
-        ''' A dictionary of sensor config. Some values can be updated, some are read-only. [dict]'''
-        data = self._get('config')
-        self._config.clear()
-        self._config.update(data)
-        return self._config
-
-    @config.setter
-    def config(self, data):
-        self._config.clear()
-        self._config.update(data)
-
-    @property
-    def recycle(self):
-        ''' True if this resource should be automatically removed when the last reference to it disappears [bool]'''
-        self._recycle = self._get('manufacturername')
-        return self._manufacturername
-
-
-class Group(Light):
-
-    """ A group of Hue lights, tracked as a group on the bridge
-
-    Example:
-
-        >>> b = Bridge()
-        >>> g1 = Group(b, 1)
-        >>> g1.hue = 50000 # all lights in that group turn blue
-        >>> g1.on = False # all will turn off
-
-        >>> g2 = Group(b, 'Kitchen')  # you can also look up groups by name
-        >>> # will raise a LookupError if the name doesn't match
-
-    """
-
-    def __init__(self, bridge, group_id):
-        Light.__init__(self, bridge, None)
-        del self.light_id  # not relevant for a group
-
-        try:
-            self.group_id = int(group_id)
-        except:
-            name = group_id
-            groups = bridge.get_group()
-            for idnumber, info in groups.items():
-                if PY3K:
-                    if info['name'] == name:
-                        self.group_id = int(idnumber)
-                        break
-                else:
-                    if info['name'] == name.decode('utf-8'):
-                        self.group_id = int(idnumber)
-                        break
-            else:
-                raise LookupError("Could not find a group by that name.")
-
-    # Wrapper functions for get/set through the bridge, adding support for
-    # remembering the transitiontime parameter if the user has set it
-    def _get(self, *args, **kwargs):
-        return self.bridge.get_group(self.group_id, *args, **kwargs)
-
-    def _set(self, *args, **kwargs):
-        # let's get basic group functionality working first before adding
-        # transition time...
-        if self.transitiontime is not None:
-            kwargs['transitiontime'] = self.transitiontime
-            logger.debug("Setting with transitiontime = {0} ds = {1} s".format(
-                self.transitiontime, float(self.transitiontime) / 10))
-
-            if (args[0] == 'on' and args[1] is False) or (
-                    kwargs.get('on', True) is False):
-                self._reset_bri_after_on = True
-        return self.bridge.set_group(self.group_id, *args, **kwargs)
-
-    @property
-    def name(self):
-        '''Get or set the name of the light group [string]'''
-        if PY3K:
-            self._name = self._get('name')
-        else:
-            self._name = self._get('name').encode('utf-8')
-        return self._name
-
-    @name.setter
-    def name(self, value):
-        old_name = self.name
-        self._name = value
-        logger.debug("Renaming light group from '{0}' to '{1}'".format(
-            old_name, value))
-        self._set('name', self._name)
-
-    @property
-    def lights(self):
-        """ Return a list of all lights in this group"""
-        # response = self.bridge.request('GET', '/api/{0}/groups/{1}'.format(self.bridge.username, self.group_id))
-        # return [Light(self.bridge, int(l)) for l in response['lights']]
-        return [Light(self.bridge, int(l)) for l in self._get('lights')]
-
-    @lights.setter
-    def lights(self, value):
-        """ Change the lights that are in this group"""
-        logger.debug("Setting lights in group {0} to {1}".format(
-            self.group_id, str(value)))
-        self._set('lights', value)
-
-
-class AllLights(Group):
-
-    """ All the Hue lights connected to your bridge
-
-    This makes use of the semi-documented feature that
-    "Group 0" of lights appears to be a group automatically
-    consisting of all lights.  This is not returned by
-    listing the groups, but is accessible if you explicitly
-    ask for group 0.
-    """
-    def __init__(self, bridge=None):
-        if bridge is None:
-            bridge = Bridge()
-        Group.__init__(self, bridge, 0)
-
-
-class Scene(object):
-    """ Container for Scene """
-
-    def __init__(self, sid, appdata=None, lastupdated=None,
-                 lights=None, locked=False, name="", owner="",
-                 picture="", recycle=False, version=0):
-        self.scene_id = sid
-        self.appdata = appdata or {}
-        self.lastupdated = lastupdated
-        if lights is not None:
-            self.lights = sorted([int(x) for x in lights])
-        else:
-            self.lights = []
-        self.locked = locked
-        self.name = name
-        self.owner = owner
-        self.picture = picture
-        self.recycle = recycle
-        self.version = version
-
-    def __repr__(self):
-        # like default python repr function, but add sensor name
-        return '<{0}.{1} id="{2}" name="{3}" lights={4}>'.format(
-            self.__class__.__module__,
-            self.__class__.__name__,
-            self.scene_id,
-            self.name,
-            self.lights)
 
 
 class Bridge(object):
@@ -610,7 +50,7 @@ class Bridge(object):
             self.config_file_path = config_file_path
         elif os.getenv(USER_HOME) is not None and os.access(os.getenv(USER_HOME), os.W_OK):
             self.config_file_path = os.path.join(os.getenv(USER_HOME), '.python_hue')
-        elif 'iPad' in platform.machine() or 'iPhone' in platform.machine() or 'iPad' in platform.machine():
+        elif MOBILE:
             self.config_file_path = os.path.join(os.getenv(USER_HOME), 'Documents', '.python_hue')
         else:
             self.config_file_path = os.path.join(os.getcwd(), '.python_hue')
@@ -1224,23 +664,3 @@ class Bridge(object):
 
     def delete_schedule(self, schedule_id):
         return self.request('DELETE', '/api/' + self.username + '/schedules/' + str(schedule_id))
-
-if __name__ == '__main__':
-    import argparse
-
-    logging.basicConfig(level=logging.DEBUG)
-
-    parser = argparse.ArgumentParser()
-    parser.add_argument('--host', required=True)
-    parser.add_argument('--config-file-path', required=False)
-    args = parser.parse_args()
-
-    while True:
-        try:
-            b = Bridge(args.host, config_file_path=args.config_file_path)
-            break
-        except PhueRegistrationException as e:
-            if PY3K:
-                input('Press button on Bridge then hit Enter to try again')
-            else:
-                raw_input('Press button on Bridge then hit Enter to try again')  # noqa
