+++ phue/exception.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+
+
+class PhueException(Exception):
+    def __init__(self, id, message):
+        self.id = id
+        self.message = message
+
+
+class PhueRegistrationException(PhueException):
+    pass
+
+
+class PhueRequestTimeout(PhueException):
+    pass
